requirements.txt
solana
python-dotenv